+++ lib/elliptic.dart
@@ -3,4 +3,8 @@
 /// More dartdocs go here.
 library elliptic;
 
+// export all
+export 'src/base.dart';
+export 'src/publickey.dart';
+export 'src/privatekey.dart';
 export 'src/elliptic.dart';