+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.2.0
+
+- Null-safe
+- Fix bugs
+- Pass tests (mainly P256)
+- Add utils to PublicKey and PrivateKey
+
 ## 0.1.0
 
 - Initial version, created by Stagehand
+- Having bugs, dont use this version